Conflicts:
	python/yt/wrapper/tests/test_native_mode.py
	scripts/reshard.py
	yt/server/cell_master/bootstrap.cpp
	yt/server/data_node/block_store.cpp
	yt/server/data_node/chunk_cache.cpp
	yt/server/exec_agent/job.cpp
	yt/server/exec_agent/slot.cpp
	yt/server/exec_agent/slot.h
	yt/server/exec_agent/slot_manager.h
	yt/server/job_proxy/user_job.cpp
	yt/server/scheduler/map_controller.cpp
	yt/server/scheduler/merge_controller.cpp
	yt/server/scheduler/operation_controller_detail.cpp
	yt/ytlib/api/client.h
	yt/ytlib/chunk_client/replication_reader.cpp
	yt/ytlib/driver/table_commands.h
	yt/ytlib/query_client/plan_fragment.proto
	yt/ytlib/table_client/pipe.cpp

+++ yt/core/compression/codec.cpp
@@ -25,7 +25,7 @@ class TCodecBase
     : public ICodec
 {
 protected:
-    static int ZeroSizeEstimator(const std::vector<int>&)
+    static size_t ZeroSizeEstimator(const std::vector<int>&)
     {
         return 0;
     }
@@ -58,7 +58,7 @@ protected:
         TConverter converter,
         bool compress,
         const std::vector<TSharedRef>& refs,
-        std::function<int(const std::vector<int>&)> outputSizeEstimator = ZeroSizeEstimator)
+        std::function<size_t(const std::vector<int>&)> outputSizeEstimator = ZeroSizeEstimator)
     {
         // XXX(sandello): Disable tracing to due excessive output.
         // auto guard = CreateTraceContextGuard(compress);
@@ -71,7 +71,7 @@ protected:
         }
 
         std::vector<int> inputSizes;
-        i64 totalInputSize = 0;
+        size_t totalInputSize = 0;
         for (const auto& ref : refs) {
             inputSizes.push_back(ref.Size());
             totalInputSize += ref.Size();
